Add `prettier-eslint` for JavaScript formatting
This adds the `prettier` pretty-printer for JavaScript to our project.
Basically, it parses JS to an AST and rewrites it in an automated way,
and then runs the output of that through `eslint --fix`, so we can set
our own preferences on top of prettier's sensible defaults.

I added a `package.json` script (`fmt`) which runs `prettier` on all the
JavaScript in `static`, except for the Flow definitions which we like to
indent for visibilities. Yay!