Refactor comments to match designs
In the mocks the 'reply' textfield for a top-level comment (i.e. a
comment replying to a post) is supposed to be always open. Currently,
it's not, and you need to click a 'reply' button to open it and start
composing a comment.

This refactors the code for rendering the commenting form, and separates
it into two separate components, one for comment replies and one for
post replies.

part of #125
pr #130