Move comment reply button to match mocks
In the mocks the 'reply' button appears to the right of the voting
buttons in the comment tree. This refactors the CreateCommentForm a
little bit so that we can render a 'reply' button (for the correct
comment) wherever we want to. There are also a few small CSS tweaks that
were necessary to keep the spacing and so on.

closes #125
pr #138