Remove error when clicking 'cancel' on create post page
We had an error on the create post page which would break the display of
posts on the channel after clicking 'submit'. What was happening was
that the button was inside of a `<form>` tag, and the form was issuing a
`POST` request, because we weren't calling `preventDefault`.  We needed
to add `.preventDefault()` on the event passed to our 'cancel' button's
`onclick` handler.

closes #151
pr #170