Get scroll behavior on page transitions to work in the normal way
This fixes the scroll position behavior on react-router route changes to
be, well, normal. The normal way that this should work is this:

- when you click on an `<a>` tag you open the new page, scrolled to the
  top, regardless of whether you've visited it before

- if you use the forward / backward buttons to navigate around, then the
  scroll position on any pages you visit should be preserved.

This adds a little component (called `ScrollToTop`) which wraps the
whole app and gives us this behavior. Basically, we can check that the
location has changed by checking that the `location` prop from the
`withRouter` HOC is different, and we can look at `history.action` to
figure out why it has changed. If the user clicked on a `<Link />` the
`history.action` will be `'PUSH'`, so in this case we can scroll to the
top, and otherwise we can allow react-routers default behavior to happen
(preserving scroll position with the backward / forward buttons).

closes #157
pr #172