Change create post form to have a channel select dropdown
This basically refactors the create post form so that we can show a
'create post' link on any page. Previously we would only show it on the
'channel' pages. Basically, this adds a dropdown with the list of the
user's subscribed channels to the page, so that the user can pick which
subreddit the post should go to. If they click 'create post' on a
subreddit that entry will be pre-selected for them. They cannot submit
the form until they pick a subreddit.

closes #126
pr #191