Fix issue w/ comment submit button being disabled during upvoting
This fixes a little UI quirk, where upvoting a comment would prevent the
user from submitting a new top-level comment or replying to an existing
comment. This uses the same pattern we're using on the upvote / downvote
buttons to get around that (using `this.state` to track whether that
particular instance was clicked on or not).

closes #212
pr #240