Add validation when creating a post and make 'title' field a textarea
This makes the title input for a Post into a `<textarea />` (instead of
`<input type="text" />`), and also writes some code for validating the
form inputs on the post create page. I wrote some general, somewhat
abstract validation code which should be useful for validating other
things as well.

closes #251
closes #243
pr #275