Add subreddit title to the API and frontend
Basically we want to show the channel title in the little byline, not
the channel name (which basically a url component). To do this I made a
small change to the channel serializer. This basically added a new
request to a ton of tests in `channels/views_tests.py` so I had to
regenerate a whole bunch of cassettes. As a nice side effect, several
tests are now converted to use the new, reproducible factory strategy
for creating reddit objects that Nathan came up with. So this should be
much less painful for the next person who has to fiddle with it, haha.

closes #257
pr #338