Set focus on comment reply forms, add key combo to submit
This sets the `autoFocus` prop on the `textarea` for replying to
comments, so that we you click the 'reply' button the focus
automatically goes into the text box (and you can compose your comment
without having to click or tab into it first).

I also added an `onKeyDown` handler to the reply form, which listens for
the ctrl-enter key combo and submits the form when you press that (for
that power user life).

closes #237
pr #291