Split the PostDisplay component into two separate components
Basically we currently have the `PostDisplay` component have two
different modes, one which is used in a list-view (frontpage, channel
view) and another which is used on a post detail view. To work on adding
post editing I need to modify the post detail version but not the
list-view version, and I think it's just too much to have it all in the
same component. So, this splits it out into two different components,
each of which is only responsible for one thing. They still share a lot
of styling, and code which is common to the two I pulled out into helper
functions in `lib/posts.js`.

closes #327
pr #331