Fix channel navigation error
This will basically clear the channels error object when trying to load
data on the ChannelPage. This fixes the bug that was occurring in the
linked issue, which was happening becuase the error from visiting a
non-existant subreddit was sticking around even after the URL had been
changed to an existing one. Since the UI just looks for the presence of
the error property to see if it should show UI or not this was
preventing the actually existing subreddit from showing up, even when
the data was coming back from the backend ok.

closes #323
pr #356