Add a 404 message to the channel page
This adds a 404 message to the channel page, which we show when we get
that error code back from the API.

This also fixes an error where, when looking at a non-existent channel
or one for which the current user does not have permission, we generated
a 500 error when trying to get the channelModerators for that channel.
We can get around that by only issuing the request to get channel
moderators after the get channel request has come back ok.

This also makes the PostPage container clean up after itself a little
bit, clearing Post error states on ComponentWillUnmount if they are
present (so they can't hang around and wreak havoc, like spooky lil
ghost errors 👻🦇)

closes #408
closes #393
pr #422