add UI for choosing post sort method
This adds UI on the home and channel pages for choosing the method used
to sort the post list. At this point we support the 'hot', 'new', and
'top' options. We store the selection in a url query param, so users can
link to a particular sort method if they want.

closes #228
pr #439