Add comment sorting UI
This adds some UI to set the 'sort' param when you're fetching the
comments for a post. Basically we add a little dropdown that has the
sorting options, and we you pick one we re-issue the request for
comments to the backend with the relevant param. The current value for
the 'select' drop down is stored as a URL query param as well, so a user
could link to a particular sort method if they wanted to.

closes #229
pr #457