Automatically render plain URLs in Markdown as <a> tags
We want to automatically render URLs as links, so the user can input a
link and have it be 'clickable' without having to know anything about
Markdown.

I looked and the most recent version of the React markdown renderer we
are using ('react-markdown') now supports doing this automatically. So I
bumped us up to the latest version, and created a new component
(Markdown) which just includes a few small tweaks to the defaults for
this package.

closes #440
pr #472