Fix non-moderator comment editing
This PR fixes an error that was occurring for non-moderator users when
editing comments. Basically, we were patching the whole comment object,
which includes fields which are read-only for non-moderators. So when
the backend tried to save the object, the user would get a 403 because
they do not have the ability to modify e.g. the 'removed' field.

This changes the comment editing so that we only post the things we need
to.

closes #476
pr #478