Fix error with non-moderators editing posts
This fixes an issue similar to the comment editing issue. We were
posting the whole Post object, but for non-moderator users there are a
few read-only fields on the object. So, when they tried to edit a post,
they would get a 403 error. Unlike the comment issue, the edit wouldn't
even go through (the comment edit would go through but then show an
error page).

Anyway, this applies the same fix I did for the comment issue, so we
should be good going forward.

closes #482
pr #488