Fix 403 error on post page (from moderator API)
This PR does 2 things:

- fixes the issue in #393, which is that, on the Post page, we were
still issuing the request to get moderators for the channel even if the
post and channel came back with 404 or 403 errors.

- I also noticed that if you visited a post on a channel which does not
exist the 404 page was not rendering (because we checked first if the
`channel` was nil), so I changed it so that we would render the 404
component in such a case.

closes #393
pr #491