Add settings page for adjusting notification prefs
This adds a page at /settings/ where the user can adjust their
notification preferences. If the user is logged in, it will issue
requests to fetch the settings data and so on in the normal way, and if
there is a stateless auth token provided in the URL it will instead use
that to authenticate (both to fetch information and to issue PATCH
requests).

closes #451
pr #511