Add error page for 403 error
This adds a 403 error page, which we show on the channel, post, and
channel moderation pages whenever the user gets a 403 (not-authorized)
error back from the API. This can happen when, for instance, the user
tried to navigate directly to a channel which they are not part of.

I also moved the 404-page handling logic into the `<Loading />` component,
so it lives in one place, instead of in several.

closes #583
pr #614