Add embedly frontend code
Now when we're going to show a link-type post on the frontend we first
hit the embedly API and use that to show some rich content from the
link. If the content is directly embeddable, we'll show it directly
inline (e.g. a YouTube video or something). If it's not directly
embeddable, we'll show the title, description, and a thumbnail image.

closes #604
pr #653