Add a link preview to the link post creation screen
This adds a little 'preview' feature to the link post creation screen.
Basically, when you add a link in the URL field, or change the URL
that's already there, we get the embed data from the embedly API and
show that inline. This should give the user a good idea of what their
finished post will end up looking like.

closes #694
part of #604
pr #685