Fix bug related to fetching subscriptions in App.js
This fixes an issue with the logic (in `App.js`) for when to fetch the
channel subscriptions. The way this affected the user was that, if you
entered the site from the `/settings` page and then navigated, without
doing a full-page refresh, from there to the homepage the channel
subscriptions would not be fetched, so there would be an empty
subscriptions box in the UI.

The issue was because we don't want to fetch the channels on
`/settings`, because on `/settings` we allow the user to authenticate
with a token as well as a normal JWT session (and the token can't
authenticate to the subscriptions API). So in `App.js` we had logic to
not fetch the subscribed channels if we were on the settings page. I
added a bit of logic which runs on `componentDidUpdate` in `App.js` and
just basically checks for the case where we've just navigated away from
the settings page and, if so, we fetch the channel subscriptions, so
that the user will see a list in the sidebar.

closes #684
pr #714