Fix issues with html returned from Embed.ly link type
This just makes the embedly rendering code a little more foolproof. It
turns out that embedly will sometimes (not realiably) return an 'html'
value on link-type posts. Previously, the code assumed that the 'html'
key would only have a usable value in it on video-type posts, so if a
link returned html it would display in a weird squished video
aspect-ratio box. This fixes that, so that we only put videos in the
`.video-container` div, and we render the `.html` content for links if
it's there.

closes #717
pr #718