Fix issue with twitter embeds
This fixes an issue which was showing up when twitter things were
embedded in link posts. The issue was that, because we were adding the
twitter embed html to the page after page-load, the twitter JS wasn't
being sourced correctly. I added some stuff to both add the twitter
JavaScript on the pages where it's necessary and also ensure that we
load the widget when needed.

closes #744
pr #745