Fix iframe styling issue
This fixes a css issue. The issue was that, in the embedly component, an
iframe has position: absolute and top: 0, left: 0 on it, which causes it
to block out the whole screen (lol). this styling is supposed to be only
for videos, where we want to put the position: absolute iframe inside of
a position: relative div, do be able to get a responsive container
for the video with consistent aspect ratio. anyway, this CSS should
be limited to just video embeds, not all embeds which have an iframe.

closes #757
pr #758