Implement WYSIWYG editor for Posts
This adds a new component, `<Editor />`, which uses ProseMirror to
implement a WYSIWYG editor to allow the user to input rich formatting.
This is serialized, by the component, to markdown. Initially, we're just
going to use this on the Post creation screen and when editing a post.

The `<Editor />` component itself it pretty straightforward to use, it
just accepts an `onChange` prop which it calls with a markdown string.

Happy markdowning!

closes #725
closes #965