Update drawer and toolbar layout!
This implements a new design for the toolbar and the drawer. The drawer
now is more of a material-style drawer, so on desktop it opens as a
little panel to the left of the main content which squeezes the main
content over, and on mobile it opens as a panel floating over the main
content.

To change the layout, I had to also change a few related layout things,
and simplify a few things had written to enable our old layout. Overall,
I think the layout should be more straightforward now, and it's nice and
responsive across all display sizes.

closes #724
pr #799