Fix silly bug with embedly display
This fixes a really silly mistake I made!

This makes sure we're dealing with a URL post before we display the
Embedly component. I recently changed the Embedly component to include a
loader placeholder. Previously, it had its own logic that would only
display when it got a result, so it was safe to just leave it on the
page at all times. After I changed it to have a 'loading' placeholder it
was no longer safe! oops.

This just checks to make sure we're dealing with a link post before we put the component on the page.

pr #817