Update post detail page to new design
This updates the display of the post to match the most recent designs.
This includes a refactor of how we display the little post action
buttons, adds a new sharing menu, and a few other things.

As part of this PR I also updated the flow-typed typedefs for all our
installed dependencies. We were missing, somewhat crucially, the
typedefs for react-redux, so none of our connected components were being
properly type-checked (because Flow didn't know how the type signature
of the `connect` function). So, now that that has been added, we should
be in a more-completely-type-checked world.

This change also involved a small refactor of the profile image
component, to add a "micro" size option.

closes #834
pr #869