Remove a flow workaround
I needed to use `React.Fragment` in a previous PR, but the version of
flow which we were then using threw an error if you just used
`<React.Fragment>` directly in JSX. To get around the error I added a
qualified import, `import React, { Fragment } from "react"`, which
silenced the error. The newer version of flow we recently upgraded to no
longer has this issue, so we can remove the workaround now.

pr #858