Fix flow issues with component prop typing
This fixes our flow typing for a number of components. Before, we had a
lot of components typed like this:

```js
class FooBar extends React.Component<*> {
  props: FooBarProps
}
```

This isn't right! If you use Flow to get the type of this class it is
`React.Component<*>` when it should be `React.Component<FooBarProps>`.
This setup will type-check the usage of props *inside* the component but
will *not* check that the right props are passed in to the component.
This means that we're effectively not using like 50% of what Flow can
give us.

We should instead be typing things like this:

```js
class FooBar extends React.Component<FooBarProps> {
}
```

this actually type-checks that the right props are passed in to
components when we use them elsewhere in the codebase.

I went through and made this change almost everywhere, there were a few
places that it ended up being too difficult to get it to type-check so I
left it alone.

pr #918