fix rich embed display width
this fixes the 'rich' type embed display a little bit. basically, this
just changes things so that, if the 'rich' type is an iframe, we do the
same responsive iframe thingy we currently do for videos. sometimes,
notably with wikipedia articles, 'rich' type embeds are not iframes, so
we don't always want to do this.

pr #984