Switch over to the material grid
This PR redoes our layout to use the material grid for column widths,
instead of our own custom stuff. I added two react components, <Grid />
and <Cell />, which are basically just divs with the right classes set
on them. These can be used to make a material grid wherever we want.  I
went ahead and replaced the existing layout in our `singleColumn` and
`withChannelSidebar` components with this setup.

closes #1007
pr #1110