create new helper function for simple component tests
We do a lot of testing of simple components using enzyme's `shallow`. A
common pattern is to declare a one-off method that takes props and
renders the component with `shallow`.

I think it's a good pattern, but we also do some other things in a few
places. This PR basically just creates a new method,
`configureShallowRenderer`, which lets you easily setup a simple test
function which works like `renderComponent` above. You call it with your
component and a default props object, and it returns a function which
accepts optional extra props and returns an enzyme shallow wrapper. This
just makes our tests a bit more standardized.

pr #1146