Install storybook and set up a few basic stories
This installs the storybook-react package and a few addons, and just
basically sets up a skeleton for using it for component development. We
may want to change how we organize things in the future, but for now
everything related to storybook lives in `static/js/storybook`. All the
stories are in an `index.js` file, and we don't really have much more
structure than that. If we end up using the tool a lot we probably want
to split out stories into separate files, and maybe colocate them with
the code.

pr #1318