Link and button styling changes
This fixes the link styling for a few little things and changes the
button style for the 'Write something' and 'Add a link' buttons on the
post creation page.

I also went ahead and added stories for links and buttons to Storybook,
so we can easily look at the different button and link styles we have
there.

closes #1314
closes #1292
closes #1313
pr #1334