Fix spacing for top of channel loading animation
This fixes a little style regression. when I made changes to add the
channel banner to the top of the channel page I also changed things to
set `margin-top: 0;` on the channel page. This had the side-effect of
getting ride of the top padding when the loader animation was playing,
so it looked a bit silly. This just adds some top padding only to the
loading channel page, so it looks nice.

pr #1339