Add URL validation to create post form, fix issue with cancel button
this fixes two issues:

- Adds URL validation to check that a URL is valid to the post create
form. We already use the same library to validate URLs before we submit
the URL to embedly.

- adds `type="button"` to the 'cancel' button on the create post form.
not having this was creating a strange issue, where hitting 'enter' in
the 'url' field would trigger the `onClick` handler on this button, so
you would go back to the previous page. Adding `type="button"` fixes
this.

these should, together, make adding a link post a little bit more
user-friendly.

closes #1366
pr #1375