actually, make the ReactNodeView constructor take a component argument
that way we can easily override the display for any of the built-in
node types that we might want to, without having to edit the schema
itself. i'm not sure we'll want to do that, but if we want to support
some types of editing for the image node inline we may want something
like that.