Add CKEditor for creating Article posts
This adds our custom build of CKEditor5 and sets us up to use it for
creating Article posts. It also sets up the EasyImage service, which we
can use to allow users to easily upload images in the article.

All in all, these changes involved the following things:

- adding the ckeditor build as a dependency
- upgrading webpack to version 4, so that it would work with the
existing ckeditor development toolchain. upgrading to webpack 4 also
required some config changes for things where the name or the
configuration format changed. one of these things was that it no longer
always spits out a common bundle, regardless of whether there is any
code in it. before we were including a common bundle on every page, but
this file was essentially empty because we only actually have one entry
point (root.js), so there is no code common to all bundles to be pulled
out. webpack 4 seems to have changed things so that this file is only
produced if it would have any application code in it.
- creating an API for retrieving an authentication token (a JWT) for the
EasyImage service
- adding environment variables for holding the EasyImage credentials
(CKEDITOR_SECRET_KEY, CKEDITOR_ENVIRONMENT_ID, CKEDITOR_UPLOAD_URL)
- adding an ARTICLE_UI feature flag which can be used to turn on the
article UI (defaults to off)
- some changes to our JS tests to support browser features that CKEditor
needs, like MutationObserver and DOMParser
- an ArticleEditor React component, which wraps the CKEditor custom
build and handles things like binding an onChange function, setting
initial data, and setting an optional read-only mode (for displaying the
articles)
- changes to the post creation form to support making an article
(hidden behind the feature flag by default)
- some basic validation for articles when creating a post

I think that's about it. Still left to do is the frontend work to add
the article type as a configurable allowed type on the channel settings
page, and also editing a saved Article-type post.

closes #1456
closes #1455
closes #1326
pr #1557