Fix stacking issue with z-index banner and compact post buttons
This changes the z-index on two things, the banner and the user menu, so
that some stacking order stuff is fixed.

Basically, before there was an issue that caused some of the buttons on
the post list view cards to show up on top of the banner (see linked
issue for a screenshot). this was easily fixed by setting a sufficiently
high `z-index` on the banner, but that then caused some issued with the
user menu (specifically, it would appear under the banner...)

So, to get around that, I also added a `z-index` to the user-menu
dropdown, so it should float above the banner.

closes #1405
pr #1487