add validation for super long text posts
This just adds some basic validation that the text post body isn't
longer than the backend allows (max of 40000 characters). I think we
shouldn't run into this too often, but this change will make it so that,
if we do, the user will see a validation message instead of just getting
an error.

closes #1469
pr #1479