Add the site name to the mobile drawer header
This just adds the site name to the mobile drawer header. I pulled out
the existing markup we were using for the site name, in the toolbar, and
added it to the HamburgerAndLogo component we already use to share code
between the mobile nav drawer and the toolbar, so we can just have it
live in one place. This required a few small CSS changes as well.

closes #1467
pr #1503