post creation page tweaks
this is some appearance tweaks for the create post page to address some
issues with the CKEditor and with the article creation UX in general.

this does these things:

- the title field is renamed to 'headline', and is styled to more
  closely resemble how it will be styled on the post detail page
- the top menu for the article editor is now 'sticky', so it should
  stick to the top of the viewport if you scroll down on a long post
- these is now a placeholder when the article editor is empty
- the article editor should focus right away when you open it
- some styling tweaks to the article editor, the size of the icons in
  the bar, etc
- when you open the article editor, the navigation drawer should close, if open

I also bumped the version of our ckeditor build to the latest one, which
allows support for setting the placeholder via CSS and fixes the issue
with the image captions.

closes #1572
part of #1583
pr #1643