ask for confirmation when the user switches post types
This adds a confirmation dialog on the create post page if the user
tries to remove the post content from a post (click the little 'x') and
they have already added some input. If they haven't made any input
(typed in a url, text post, or article) we don't put up the
confirmation.

part of #1583
pr #1601