Add cover image to article post
This PR implements an initial version of the cover image for the article
post type, which lets you upload a photo to use with the article type
post.

Basically, there is a new input on the create post post, which we show
alongside the article editor, which lets the user select an image to
upload. I also added stuff to have some basic validation (just that the
thing is an image and not, say, a PDF).

closes #1594
pr #1660