delete the icons from the post type buttons
This makes a few adjustments to the post type buttons on the post create
form to make them fit a little better on mobile (especially when the
article UI flag is turned on).

Basically before, with the article post type button, some of the buttons
could get squashed too much on a phone, so that the text in them would
be forced on to two lines. This adjusts the padding and font size so
that doesn't happen any longer.

closes #1644
pr #1681