refactor API library file to several modules
This starts the work of splitting up the lib/api.js file into several
modules, all of which now live in lib/api/. I also split up the tests
that were specific to each module.

There are still some things we can def split out, like the functions for
hitting the settings APIs and the comment-related functions.

part of #1667
pr #1668