Fix a flakey JS test
fixes a flakey thing in a JS test. basically when there is a value of -0
`JSON.stringify` will turn that into just 0, which is sensible. However,
it seems that sometimes our post factories set a value of -0 for a
score. this was causing a flaky test on something I changed for #1660
(see this test failure for an example:
https://travis-ci.org/mitodl/open-discussions/jobs/477491161#L2383).

pr #1671