use post_type to show UI specific to different post types
Basically this just uses the relatively-new `post_type` value to decide
what UI to show or which logic to use in a few places where we need to
do different things based on the post type. For instance, when editing a
post we need to display either the normal text editor or the article
editor, depending on what kind of post we're dealing with. Previously we
inspected to post object to see what fields were defined on it, and from
there we chose which UI we wanted. This instead just inspects the
`post_type` value, which is both more reliable and more future-proof.

closes #1652
pr #1723