Fix a post page form reset bug
This fixes an issue with the form reset logic we have on the create post
page. Basically, there are a few cases where, when the user switches
between channels, we want to reset the input values for the form, either
because the current input isn't valid on the new channel or because the
new channel offers the user an option which isn't present on the
previous channel.

The particular issue this is fixing is a situation where we the form to
be cleared and it isn't. The situation is:

- the user navigates to `/create_post/`
- the user clicks on 'Write an Article'
- the user fills in the article and title
- the user selects a channel which allows text and link posts, but not
  article posts
- the user submits the form

On `master` right now this will fail to clear the article post type and
the article input from the form, and the user will get backend error
because the post type is not supported by the channel they've selected.
We should be clearing the form of input if the user makes a selection
like this. the problem is that the logic we have wasn't fully updated to
support three post types, but implicitly assumes that the max is two.

Anyhow, I rewrote the `shouldFormReset` function to hopefully both make
it more clear and straightforward and also to take account of this case.

I also discovered that many tests for the create post page were not
running because of a typo (i.e. 1 out of 17 tests cases was being run).
So I fixed that, and added some flow types which would have caught that.

closes #1687
pr #1690