update prosemirror-markdown to latest version
this fixes an edge case which we noticed where it was possible to create
certain markdown strings with the editor which it couldn't properly
reinitialize. see the issue for the example.

anyhow, it seems that what we noticed was an upstream bug, and bumping
to the latest version fixes it.

closes #1696
pr #1726