fix issue with post delete dialog staying open
This fixes an issue where we were not closing the post delete dialog
when we should have been. Basically, the confirm callback passed in to a
dialog didn't dispatch an action to hide the dialog after the API
request came back successfully, so the boolean flag was still set in the
redux store to show the dialog. This meant that on the next visit to a
post detail page the dialog would still show up. Even worse, confirming
that dialog would delete the post that the user was now on, a
potentially uninintended (and destructive!) action.

Anyhow, it's now fixed! 🎉🎆

closes #1686
pr #1742