Add support for rending content using Embedly in the article editor
This adds support for using embedly to display media embeds instead of
CKEditor's built-in support (which supports only a very limited number
of sites, like youtube and so on). I did this by changing the
configuration for embedly's built-in MediaEmbed thing.

Basically now any site that is supported by embedly should work just
fine. You can either add media embeds to the editor by using the button
in the paragraph menu, or you can just paste in a link as well.

closes #1457
pr #1757