upgrade flow to @latest
I was working on #1299 and I ran into an issue where a newer feature of
React wasn't supported by the version of flow we are currently using, so
I went ahead and upgraded it.

The newer version had a big change (see here:
https://medium.com/flow-type/asking-for-required-annotations-64d4f9c1edf8)
where in many cases type parameters that should have been required were
not, and lots of typings were silently failing.

Upgrading lets us also use newer libdefs for things like `react-redux` -
in particular, with the new version of that `connect` is no longer a
type black hole!

Now instead of just having the return value of `connect` be more or less
totally untyped (as far as flow knows anyhow) it accepts a series of
parameters, and it will actually check that props that the component is
supposed to receive from parents are passed down.

To do this you need to declare new types, `StateProps`, `OwnProps`, and
`DispatchProps`. `StateProps` is the return value of the
`mapStateToProps` function, and includes any props that come from the
redux state. `OwnProps` is props which the component receives from it's
parents, and `DispatchProps` are the props which come from
`mapDispatchToProps`. Then the type which the unconnected component
itself uses is

```js
type Props = {|
  ...OwnProps,
  ...StateProps,
  ...DispatchProps
|}
```

then you annotate `connect` like

```js
export default connect<Props, OwnProps, _, DispatchProps, _,_>(mapStateToProps)(Component)
```

the `_` syntax tells Flow to infer those types (you can read the
react-redux typedef docs to see more).

Anyway, this is really nice! Unfortunately it still isn't fully
supported in the places where we use `R.compose` for this purpose, since
the ramda libdef isn't compatible easily with our usage (I'm talking
7000 type errors, lol), but for more simple components it's a good thing
imo.

pr #1904