Add some scrolling to post create page to make errors visible
This adds a little bit of scrolling to the create post page. Basically
the issue is that, if you're editing a very long post or you're working
on a short screen, you can scroll down so that the title or channel
validation errors are not visible. It won't be clear to the user in that
case why the post isn't going through.

What I decided to do to fix the issue is to check when the user submits
the post if there is a `title` or `channel` validation error and, if so,
do a smooth scroll to the top. This should only be minimally disruptive
in cases where you can already see the validation message.

closes #1796
pr #1853