tweak post pinning so that the UI fully reflects the newly pinned post
This PR evicts the current post list after the user has made a request
to toggle the pinned state of a post. This will force the UI to show the
'true' state of the post list after the pinned state is updated.

There was a difficulty before because of some of the logic relating to
the pagination. This sidesteps modifying or changing that logic by just
clearing the whole post list, so we use what we get back from the API
directly, instead of trying to figure out in which cases we want to
merge the new post list into the existing one and in which cases we want
to overwrite it.

closes #1841
pr #1882