get rid of the docker setup for JavaScript tests on travis
Before we had a system of building docker images, pushing them to docker
hub, and using those to cache dependencies and some environment setup
between our travis runs. This wasn't a great system though overall, and
wasn't really necessary.

This commit changes things so that we instead dispense with the docker
containers and just run the JS tests directly in travis' JS environment.
This lets us take advantage of the dependency cacing and so on that
travis already has between runs, and avoids the overhead of pulling down
the docker image.

pr #1884