kill some sluggishness with the ArticleEditor
Just a little tweak that occurred to me (which I probably should have
done before!).

The `ArticleEditor` component can get a little bit heavy sometimes and
lead to some noticeable UI sluggishness when editing, I think because
the `editor.getData()` method is a little bit expensive, and in our
current setup we call that method on each keystroke. I just added the
lodash `debounce` function to reduce that to once every 250ms, which
makes it feel a lot more responsive to me.

pr #1912