Add cover image to the course index page
this adds a cover image to the top of the course index page. for now it
is just showing the fallback background color and gradient that we show
on the channel page when the no channel cover image has been set. in the
future when we decide on an image we want to show we can just add it in.

to implement this I refactored the way we were showing the banner image
on the channel pages into a series of reusable components, so if we need
to add a similar cover image to another page in the future it should be
relatively straightforward.

closes #1800
pr #1934