Add 'members' page for showing the people who are members of a channel
This adds a new 'members' page to the channel router which lets you see
the members of a given channel. There is a new option on the 'people
widget' which will allow you do choose whether or not to add a link from
the people widget to the members page. I also made a few tweaks to the
people widget in the course of working on this.

closes #1932
pr #1936