Added a redirect rule to handle traffic to the lemelsonx subdomain
We need to be able to handle a redirect from the lemelsonx.mit.edu domain to a specific channel in open so this redirect will serve that purpose. We will update the DNS to be a cname of open.mit.edu so that anyone who goes to that endpoint will wind up where we want them to.