add sorting to the channel members page
this adds options to the channel members page to sort the channel
members. right now we support two fields:

- sorting by name, alphabetical
- sorting by channel join date, with the most recent at the top

both of these sorting options required some changes to the index. for
the former, I had to convert `author_name` on the `profile` index from
just a simple text field to a nested field with a text field for
searching and a keyword field for sorting.

for the channel join date I added a new nested object which includes
information about the channels that the user is a member of. this
includes the channel name and the date, which allows me to sort on the
channel join date by using a nested filter to filter for just the
channel join date for the channel we're currently focused on. whew!

anyhow, it all seems to work fairly well. on the frontend some changes
were necessary to our query building code, and to the UI for the members
page (mainly to add the dropdown for the sort options).

closes #1941
pr #1956